MTN #5 rename app_setting user_setting
